+++ .storybook/webpack.config.js
@@ -2,8 +2,6 @@
 /* global __dirname */
 const path = require("path");
 
-console.log(path.resolve("src"))
-
 module.exports = {
   resolve: {
     modules: [
@@ -14,12 +12,4 @@ module.exports = {
       "@selfpub-ui": path.resolve("src"),
     },
   },
-  module: {
-    rules: [
-      {
-        test: /\.(woff2?|ttf|eot|svg|jpe?g|png|gif)$/,
-        use: ["file-loader?name=[path][name].[hash:8].[ext]"],
-      },
-    ]
-  }
 };