+++ src/assets/index.js
@@ -1,0 +1,1 @@
+export { iconsSprite, openBookIcon } from "./icons-string";