+++ src/components/index.js
@@ -1,15 +1,20 @@
 /* global process */
 export { default as Button } from "./button/index";
 export { default as Checkbox } from "./checkbox/index";
+export { default as Radio } from "./radio/index";
+export { default as RadioGroup } from "./radio-group/index";
+export { default as InfoBox } from "./info-box/index";
 export { default as FieldGroup } from "./field-group/index";
 export { default as Icon } from "./icons/index";
 export { default as InlineSelect } from "./inline-select/index";
+export { default as Select } from "./select/index";
 export { default as Link } from "./link/index";
 export { default as Loader } from "./loader/loader";
 export { default as Popover } from "./popover/index";
 export { default as Textarea } from "./textarea/index";
-export { default as ThemeProviderWrapper } from "./theme/index";
+export { ThemeProviderWrapper, colors } from "./theme/index";
 export { Typography, Header, Paragraph } from "./typography/index";
+export { Toggle, ToggleOption } from "./toggle/index";
 export {
   Input,
   InputStatus,
@@ -20,4 +25,3 @@ export {
 export { Table, Td, Tr } from "./table/index";
 export { Tab, Tabs } from "./tabs/index";
 export { default as SearchBox } from "./search-box/index";
-export { colors } from "./theme/index";